refactor: fix TypeError in SQLSRV/Connection.php
3) CodeIgniter\Database\Live\FabricatorLiveTest::testCreateAddsToDatabase
TypeError: CodeIgniter\Database\SQLSRV\Connection::insertID(): Return value must be of type int, string returned

/home/runner/work/CodeIgniter4/CodeIgniter4/system/Database/SQLSRV/Connection.php:195
/home/runner/work/CodeIgniter4/CodeIgniter4/system/Model.php:328
/home/runner/work/CodeIgniter4/CodeIgniter4/system/BaseModel.php:782
/home/runner/work/CodeIgniter4/CodeIgniter4/system/Model.php:707
/home/runner/work/CodeIgniter4/CodeIgniter4/system/Test/Fabricator.php:475
/home/runner/work/CodeIgniter4/CodeIgniter4/tests/system/Database/Live/FabricatorLiveTest.php:41

+++ system/Database/SQLSRV/Connection.php
@@ -190,7 +190,7 @@ class Connection extends BaseConnection
      */
     public function insertID(): int
     {
-        return $this->query('SELECT SCOPE_IDENTITY() AS insert_id')->getRow()->insert_id ?? 0;
+        return (int) ($this->query('SELECT SCOPE_IDENTITY() AS insert_id')->getRow()->insert_id ?? 0);
     }
 
     /**
